# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: N-Tech C³ Engineering Intelligence Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # Stories
  /stories:
    get:
      operationId: listStories
      tags: [stories]
      summary: List stories
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: campaignId
          in: query
          schema:
            type: ["number", "null"]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of stories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Story"
    post:
      operationId: createStory
      tags: [stories]
      summary: Create a story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
  /stories/{id}:
    get:
      operationId: getStory
      tags: [stories]
      summary: Get a story
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Story
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
        "404":
          description: Not found
    patch:
      operationId: updateStory
      tags: [stories]
      summary: Update a story
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoryPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
        "404":
          description: Not found
    delete:
      operationId: deleteStory
      tags: [stories]
      summary: Delete a story
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List templates
      parameters:
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "404":
          description: Not found
    patch:
      operationId: updateTemplate
      tags: [templates]
      summary: Update a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplatePatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "404":
          description: Not found
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
components:
  schemas:
...
    Project:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        color:
          type: string
    ProjectPatch:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]

    # Stories
    Story:
      type: object
      required: [id, title, status, createdAt, updatedAt]
      properties:
        id:
          type: number
        title:
          type: string
        slug:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        status:
          type: string
          enum: [Research, Idea, Draft, Review, Approved, Published, Archived]
        category:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
          enum: [Low, Medium, High, Critical, null]
        audience:
          type: ["string", "null"]
        difficulty:
          type: ["string", "null"]
          enum: [Beginner, Intermediate, Advanced, null]
        tags:
          type: array
          items:
            type: string
        projectId:
          type: ["number", "null"]
        campaignId:
          type: ["number", "null"]
        evidenceScore:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    StoryInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        summary:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [Research, Idea, Draft, Review, Approved, Published, Archived]
        category:
          type: string
        priority:
          type: string
          enum: [Low, Medium, High, Critical]
        audience:
          type: string
        difficulty:
          type: string
          enum: [Beginner, Intermediate, Advanced]
        tags:
          type: array
          items:
            type: string
        projectId:
          type: ["number", "null"]
        campaignId:
          type: ["number", "null"]
    StoryPatch:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        summary:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        status:
          type: string
          enum: [Research, Idea, Draft, Review, Approved, Published, Archived]
        category:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
          enum: [Low, Medium, High, Critical, null]
        audience:
          type: ["string", "null"]
        difficulty:
          type: ["string", "null"]
          enum: [Beginner, Intermediate, Advanced, null]
        tags:
          type: array
          items:
            type: string
        projectId:
          type: ["number", "null"]
        campaignId:
          type: ["number", "null"]

    # Campaigns
...
    Template:
      type: object
      required: [id, title, type, createdAt, updatedAt]
      properties:
        id:
          type: number
        title:
          type: string
        type:
          type: string
          enum: [LinkedIn, Blog, XThread, Facebook, ReleaseNotes, CaseStudy, EngineeringJournal, RepositoryAudit, ArchitectureReview, SprintReview, Retrospective, Other]
        content:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TemplateInput:
      type: object
      required: [title, type]
      properties:
        title:
          type: string
          minLength: 1
        type:
          type: string
          enum: [LinkedIn, Blog, XThread, Facebook, ReleaseNotes, CaseStudy, EngineeringJournal, RepositoryAudit, ArchitectureReview, SprintReview, Retrospective, Other]
        content:
          type: string
        description:
          type: string
    TemplatePatch:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        type:
          type: string
          enum: [LinkedIn, Blog, XThread, Facebook, ReleaseNotes, CaseStudy, EngineeringJournal, RepositoryAudit, ArchitectureReview, SprintReview, Retrospective, Other]
        content:
          type: ["string", "null"]
        description:
          type: ["string", "null"]