# lib/api-zod/src/generated/types/templatePatch.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * N-Tech C³ Engineering Intelligence Operating System API
 * OpenAPI spec version: 0.1.0
 */
import type { TemplatePatchType } from './templatePatchType';

export interface TemplatePatch {
  /** @minLength 1 */
  title?: string;
  type?: TemplatePatchType;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  description?: string | null;
}